+++ app/Http/Controllers/ShoppingLocationController.php
@@ -21,11 +21,13 @@ class ShoppingLocationController extends Controller
 
     public function create()
     {
+        abort_unless(auth()->user()->can('create shopping locations'), 403);
         return Inertia::render('Master/ShoppingLocations/Create');
     }
 
     public function store(Request $request)
     {
+        abort_unless(auth()->user()->can('create shopping locations'), 403);
         $validated = $request->validate([
             'name' => 'required|string|max:100|unique:shopping_locations',
         ]);
@@ -37,6 +39,7 @@ class ShoppingLocationController extends Controller
 
     public function edit(ShoppingLocation $shoppingLocation)
     {
+        abort_unless(auth()->user()->can('edit shopping locations'), 403);
         return Inertia::render('Master/ShoppingLocations/Edit', [
             'location' => $shoppingLocation,
         ]);
@@ -44,6 +47,7 @@ class ShoppingLocationController extends Controller
 
     public function update(Request $request, ShoppingLocation $shoppingLocation)
     {
+        abort_unless(auth()->user()->can('edit shopping locations'), 403);
         $validated = $request->validate([
             'name' => 'required|string|max:100|unique:shopping_locations,name,' . $shoppingLocation->id,
         ]);
@@ -55,6 +59,7 @@ class ShoppingLocationController extends Controller
 
     public function destroy(ShoppingLocation $shoppingLocation)
     {
+        abort_unless(auth()->user()->can('delete shopping locations'), 403);
         $shoppingLocation->delete();
 
         return redirect()->route('shopping-locations.index')->with('success', 'Lokasi tujuan berhasil dihapus.');
